create main function in xlsx_reader.py and added test files
